+++ test/src/node.rs
@@ -8,7 +8,7 @@ use ckb_chain_spec::ChainSpec;
 use ckb_error::AnyError;
 use ckb_jsonrpc_types::{BlockFilter, BlockTemplate, TxPoolInfo};
 use ckb_jsonrpc_types::{PoolTxDetailInfo, TxStatus};
-use ckb_logger::{debug, error};
+use ckb_logger::{debug, error, info};
 use ckb_resource::Resource;
 use ckb_types::{
     bytes,
@@ -27,6 +27,9 @@ use std::path::PathBuf;
 use std::process::{Child, Command, Stdio};
 use std::thread::sleep;
 use std::time::{Duration, Instant};
+
+#[cfg(target_os = "windows")]
+use windows_sys::Win32::System::Console::{GenerateConsoleCtrlEvent, CTRL_C_EVENT};
 
 struct ProcessGuard {
     pub name: String,
@@ -64,7 +67,7 @@ impl Node {
 
         // Copy node template into node's working directory
         let cells_dir = working_dir.join("specs").join("cells");
-        ckb_logger::info!("working_dir {:?}", working_dir);
+        info!("working_dir {:?}", working_dir);
 
         fs::create_dir_all(cells_dir).expect("create node's dir");
         for file in &[
@@ -690,20 +693,38 @@ impl Node {
         drop(self.guard.take())
     }
 
-    #[cfg(not(target_os = "windows"))]
     pub fn stop_gracefully(&mut self) {
         if let Some(mut guard) = self.guard.take() {
             if !guard.killed {
-                // send SIGINT to the child
-                nix::sys::signal::kill(
-                    nix::unistd::Pid::from_raw(guard.child.id() as i32),
-                    nix::sys::signal::Signal::SIGINT,
-                )
-                .expect("cannot send ctrl-c");
+                // on nix: send SIGINT to the child
+                // on windows: use taskkill to kill the child gracefully
+                Self::kill_gracefully(guard.child.id());
                 let _ = guard.child.wait();
                 guard.killed = true;
             }
         }
+    }
+
+    #[cfg(target_os = "windows")]
+    fn kill_gracefully(pid: u32) {
+        unsafe {
+            let ret = GenerateConsoleCtrlEvent(CTRL_C_EVENT, pid);
+            if ret == 0 {
+                let err = std::io::Error::last_os_error();
+                error!("GenerateConsoleCtrlEvent failed: {}", err);
+            } else {
+                info!("GenerateConsoleCtrlEvent success");
+            }
+        }
+    }
+
+    #[cfg(not(target_os = "windows"))]
+    fn kill_gracefully(pid: u32) {
+        nix::sys::signal::kill(
+            nix::unistd::Pid::from_raw(pid as i32),
+            nix::sys::signal::Signal::SIGINT,
+        )
+        .expect("cannot send ctrl-c");
     }
 
     pub fn export(&self, target: String) {