+++ tx-pool/src/process.rs
@@ -80,9 +80,9 @@ impl TxPoolService {
         }
     }
 
-    pub(crate) async fn fetch_tx_verify_cache(&self, hash: &Byte32) -> Option<CacheEntry> {
+    pub(crate) async fn fetch_tx_verify_cache(&self, tx: &TransactionView) -> Option<CacheEntry> {
         let guard = self.txs_verify_cache.read().await;
-        guard.peek(hash).cloned()
+        guard.peek(&tx.witness_hash()).cloned()
     }
 
     async fn fetch_txs_verify_cache(
@@ -91,8 +91,11 @@ impl TxPoolService {
     ) -> HashMap<Byte32, CacheEntry> {
         let guard = self.txs_verify_cache.read().await;
         txs.filter_map(|tx| {
-            let hash = tx.hash();
-            guard.peek(&hash).cloned().map(|value| (hash, value))
+            let wtx_hash = tx.witness_hash();
+            guard
+                .peek(&wtx_hash)
+                .cloned()
+                .map(|value| (wtx_hash, value))
         })
         .collect()
     }
@@ -165,7 +168,11 @@ impl TxPoolService {
                     self.callbacks.call_reject(tx_pool, &evict, reject);
                 }
                 tx_pool.remove_conflict(&entry.proposal_short_id());
-
+                // in a corner case, a tx with lower fee rate may be rejected immediately
+                // after inserting into pool, return proper reject error here
+                tx_pool
+                    .limit_size(&self.callbacks, Some(&entry.proposal_short_id()))
+                    .map_or(Ok(()), Err)?;
                 Ok(())
             })
             .await;
@@ -333,6 +340,21 @@ impl TxPoolService {
         } else {
             Ok(())
         }
+    }
+
+    pub(crate) async fn test_accept_tx(&self, tx: TransactionView) -> Result<Completed, Reject> {
+        // non contextual verify first
+        self.non_contextual_verify(&tx, None)?;
+
+        if self.chunk_contains(&tx).await {
+            return Err(Reject::Duplicated(tx.hash()));
+        }
+
+        if self.orphan_contains(&tx).await {
+            debug!("reject tx {} already in orphan pool", tx.hash());
+            return Err(Reject::Duplicated(tx.hash()));
+        }
+        self._test_accept_tx(tx.clone()).await
     }
 
     pub(crate) async fn process_tx(
@@ -659,7 +681,7 @@ impl TxPoolService {
         remote: Option<(Cycle, PeerIndex)>,
     ) -> Option<(Result<ProcessResult, Reject>, Arc<Snapshot>)> {
         let limit_cycles = self.tx_pool_config.max_tx_verify_cycles;
-        let tx_hash = tx.hash();
+        let wtx_hash = tx.witness_hash();
 
         let (ret, snapshot) = self.pre_check(&tx).await;
         let (tip_hash, rtx, status, fee, tx_size) = try_or_return_with_snapshot!(ret, snapshot);
@@ -672,7 +694,7 @@ impl TxPoolService {
             return None;
         }
 
-        let cached = self.fetch_tx_verify_cache(&tx_hash).await;
+        let cached = self.fetch_tx_verify_cache(&tx).await;
         let tip_header = snapshot.tip_header();
         let tx_env = Arc::new(status.with_env(tip_header));
 
@@ -764,7 +786,7 @@ impl TxPoolService {
             let txs_verify_cache = Arc::clone(&self.txs_verify_cache);
             tokio::spawn(async move {
                 let mut guard = txs_verify_cache.write().await;
-                guard.put(tx_hash, CacheEntry::Completed(completed));
+                guard.put(wtx_hash, CacheEntry::Completed(completed));
             });
         }
 
@@ -781,11 +803,11 @@ impl TxPoolService {
         cached: CacheEntry,
         remote: Option<(Cycle, PeerIndex)>,
     ) -> Result<(), Reject> {
-        let tx_hash = tx.hash();
+        let wtx_hash = tx.witness_hash();
         let mut chunk = self.chunk.write().await;
         if chunk.add_tx(tx, remote) {
             let mut guard = self.txs_verify_cache.write().await;
-            guard.put(tx_hash, cached);
+            guard.put(wtx_hash, cached);
         }
 
         Ok(())
@@ -796,7 +818,7 @@ impl TxPoolService {
         tx: TransactionView,
         declared_cycles: Option<Cycle>,
     ) -> Option<(Result<Completed, Reject>, Arc<Snapshot>)> {
-        let tx_hash = tx.hash();
+        let wtx_hash = tx.witness_hash();
 
         let (ret, snapshot) = self.pre_check(&tx).await;
 
@@ -810,7 +832,7 @@ impl TxPoolService {
             return None;
         }
 
-        let verify_cache = self.fetch_tx_verify_cache(&tx_hash).await;
+        let verify_cache = self.fetch_tx_verify_cache(&tx).await;
         let max_cycles = declared_cycles.unwrap_or_else(|| self.consensus.max_block_cycles());
         let tip_header = snapshot.tip_header();
         let tx_env = Arc::new(status.with_env(tip_header));
@@ -846,11 +868,32 @@ impl TxPoolService {
             let txs_verify_cache = Arc::clone(&self.txs_verify_cache);
             tokio::spawn(async move {
                 let mut guard = txs_verify_cache.write().await;
-                guard.put(tx_hash, CacheEntry::Completed(verified));
+                guard.put(wtx_hash, CacheEntry::Completed(verified));
             });
         }
 
         Some((Ok(verified), submit_snapshot))
+    }
+
+    pub(crate) async fn _test_accept_tx(&self, tx: TransactionView) -> Result<Completed, Reject> {
+        let (pre_check_ret, snapshot) = self.pre_check(&tx).await;
+
+        let (_tip_hash, rtx, status, _fee, _tx_size) = pre_check_ret?;
+
+        // skip check the delay window
+
+        let verify_cache = self.fetch_tx_verify_cache(&tx).await;
+        let max_cycles = self.consensus.max_block_cycles();
+        let tip_header = snapshot.tip_header();
+        let tx_env = Arc::new(status.with_env(tip_header));
+
+        verify_rtx(
+            Arc::clone(&snapshot),
+            Arc::clone(&rtx),
+            tx_env,
+            &verify_cache,
+            max_cycles,
+        )
     }
 
     pub(crate) async fn update_tx_pool_for_reorg(
@@ -1214,7 +1257,7 @@ fn _update_tx_pool_for_reorg(
     tx_pool.remove_expired(callbacks);
 
     // Remove transactions from the pool until its size <= size_limit.
-    tx_pool.limit_size(callbacks);
+    let _ = tx_pool.limit_size(callbacks, None);
 }
 
 pub fn all_inputs_is_unknown(snapshot: &Snapshot, tx: &TransactionView) -> bool {