+++ util/constant/src/lib.rs
@@ -4,7 +4,12 @@
 pub mod consensus;
 /// hardfork constant
 pub mod hardfork;
+/// softfork constant
+pub mod softfork;
 /// store constant
 pub mod store;
 /// sync constant
 pub mod sync;
+
+/// hardcoded default assume valid target
+pub mod default_assume_valid_target;